Finishing Player Controller
