+++ Assets/2D sprites/PLAYER/Animations/Jump.anim
@@ -21,21 +21,23 @@ AnimationClip:
   - serializedVersion: 2
     curve:
     - time: 0
-      value: {fileID: -2254625276829889376, guid: fb2066e15e1fa0a46a1f9636743e07f2, type: 3}
+      value: {fileID: -2254625276829889376, guid: 74f2576b5565fe442ba5057ed16db319, type: 3}
     - time: 0.083333336
-      value: {fileID: -8339041930550725910, guid: fb2066e15e1fa0a46a1f9636743e07f2, type: 3}
+      value: {fileID: -8339041930550725910, guid: 74f2576b5565fe442ba5057ed16db319, type: 3}
     - time: 0.16666667
-      value: {fileID: -4970676776471875818, guid: fb2066e15e1fa0a46a1f9636743e07f2, type: 3}
+      value: {fileID: -4970676776471875818, guid: 74f2576b5565fe442ba5057ed16db319, type: 3}
     - time: 0.25
-      value: {fileID: -3875430467764951074, guid: fb2066e15e1fa0a46a1f9636743e07f2, type: 3}
-    - time: 0.33333334
-      value: {fileID: -6949676248186077260, guid: fb2066e15e1fa0a46a1f9636743e07f2, type: 3}
-    - time: 0.41666666
-      value: {fileID: 1325182884686347552, guid: fb2066e15e1fa0a46a1f9636743e07f2, type: 3}
-    - time: 0.45
-      value: {fileID: 5033558275378456620, guid: fb2066e15e1fa0a46a1f9636743e07f2, type: 3}
-    - time: 0.5
-      value: {fileID: 603814315040073299, guid: fb2066e15e1fa0a46a1f9636743e07f2, type: 3}
+      value: {fileID: -3875430467764951074, guid: 74f2576b5565fe442ba5057ed16db319, type: 3}
+    - time: 0.38333333
+      value: {fileID: -3875430467764951074, guid: 74f2576b5565fe442ba5057ed16db319, type: 3}
+    - time: 0.51666665
+      value: {fileID: -6949676248186077260, guid: 74f2576b5565fe442ba5057ed16db319, type: 3}
+    - time: 0.56666666
+      value: {fileID: 1325182884686347552, guid: 74f2576b5565fe442ba5057ed16db319, type: 3}
+    - time: 0.6166667
+      value: {fileID: 5033558275378456620, guid: 74f2576b5565fe442ba5057ed16db319, type: 3}
+    - time: 0.6666667
+      value: {fileID: 603814315040073299, guid: 74f2576b5565fe442ba5057ed16db319, type: 3}
     attribute: m_Sprite
     path: 
     classID: 212
@@ -58,20 +60,21 @@ AnimationClip:
       isIntCurve: 0
       isSerializeReferenceCurve: 0
     pptrCurveMapping:
-    - {fileID: -2254625276829889376, guid: fb2066e15e1fa0a46a1f9636743e07f2, type: 3}
-    - {fileID: -8339041930550725910, guid: fb2066e15e1fa0a46a1f9636743e07f2, type: 3}
-    - {fileID: -4970676776471875818, guid: fb2066e15e1fa0a46a1f9636743e07f2, type: 3}
-    - {fileID: -3875430467764951074, guid: fb2066e15e1fa0a46a1f9636743e07f2, type: 3}
-    - {fileID: -6949676248186077260, guid: fb2066e15e1fa0a46a1f9636743e07f2, type: 3}
-    - {fileID: 1325182884686347552, guid: fb2066e15e1fa0a46a1f9636743e07f2, type: 3}
-    - {fileID: 5033558275378456620, guid: fb2066e15e1fa0a46a1f9636743e07f2, type: 3}
-    - {fileID: 603814315040073299, guid: fb2066e15e1fa0a46a1f9636743e07f2, type: 3}
+    - {fileID: -2254625276829889376, guid: 74f2576b5565fe442ba5057ed16db319, type: 3}
+    - {fileID: -8339041930550725910, guid: 74f2576b5565fe442ba5057ed16db319, type: 3}
+    - {fileID: -4970676776471875818, guid: 74f2576b5565fe442ba5057ed16db319, type: 3}
+    - {fileID: -3875430467764951074, guid: 74f2576b5565fe442ba5057ed16db319, type: 3}
+    - {fileID: -3875430467764951074, guid: 74f2576b5565fe442ba5057ed16db319, type: 3}
+    - {fileID: -6949676248186077260, guid: 74f2576b5565fe442ba5057ed16db319, type: 3}
+    - {fileID: 1325182884686347552, guid: 74f2576b5565fe442ba5057ed16db319, type: 3}
+    - {fileID: 5033558275378456620, guid: 74f2576b5565fe442ba5057ed16db319, type: 3}
+    - {fileID: 603814315040073299, guid: 74f2576b5565fe442ba5057ed16db319, type: 3}
   m_AnimationClipSettings:
     serializedVersion: 2
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 0.51666665
+    m_StopTime: 0.68333334
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
