+++ Assets/2D sprites/PLAYER/Animations/Walk.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a36b416257ea2a840a03dcd40528d066
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
