+++ Assets/Prefabs/Player.prefab
@@ -1,36 +1,5 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &816384986442352162
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 5832845381631469005}
-  m_Layer: 2
-  m_Name: SlopeCheckRear
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!4 &5832845381631469005
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 816384986442352162}
-  serializedVersion: 2
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -0.41, y: -1.002, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children: []
-  m_Father: {fileID: 7856122199337316976}
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1705551973750708276
 GameObject:
   m_ObjectHideFlags: 0
@@ -66,8 +35,6 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 1914395793123432756}
-  - {fileID: 8146929624418552425}
-  - {fileID: 5832845381631469005}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &457728839589209139
@@ -82,49 +49,42 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b9ef151f0c6e4bd418457271a5976f5d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  playerRB: {fileID: 0}
-  playerCollider: {fileID: 0}
+  rb: {fileID: 2903565161919527520}
+  boxCollider: {fileID: 8057494995023221748}
   groundCheck: {fileID: 1914395793123432756}
-  slopeCheckFront: {fileID: 8146929624418552425}
-  slopeCheckRear: {fileID: 5832845381631469005}
+  animator: {fileID: 7642683187360721432}
+  playerVelocity: {x: 0, y: 0}
+  maxSpeed: 5
+  moveSpeed: 5
+  jumpSpeed: 15
+  dashSpeed: 30
+  wallSlidingSpeed: 0.3
+  dashTime: 0.3
+  inputHoldTime: 0.1
+  wallJumpTime: 0.2
+  airFriction: 0.05
   maxHealth: 3
   currentHealth: 3
-  maxJumpCount: 1
-  playerSpeed: 10
-  maxSpeed: 6
-  jumpForce: 15
-  dashSpeed: 15
-  dashTime: 0
-  maxDashTime: 0.4
-  airFrictionMultiplier: 0.05
+  maxJumps: 1
   groundCheckRadius: 0.02
-  slopeCheckRadius: 0.24
-  pushValue: 2
-  jumpsLeft: 0
-  slopeAngle: 0
-  maxSlopeAngle: 70
-  minSlopeAngle: 10
-  fallMultiplier: 1.25
-  _debug: 1
-  grounded: 1
-  canDash: 0
-  dashing: 0
-  playerHasControl: 1
+  wallCheckRadius: 0.55
+  grounded: 0
   isOnSlope: 0
-  ascendingSlope: 0
-  descendingSlope: 0
-  dashUnlocked: 0
-  runUnlocked: 0
-  jumpUnlocked: 1
-  walkUnlocked: 0
+  isOnWall: 0
+  isDashing: 0
+  isJumping: 0
+  DEBUG: 0
+  playerHasControl: 1
+  playerWantsToJump: 0
   groundLayer:
     serializedVersion: 2
     m_Bits: 8
-  enemyLayer:
-    serializedVersion: 2
-    m_Bits: 0
   boxSize: {x: 1.03, y: 0.03}
-  PlayerSpeedX: {x: 0, y: 0}
+  moveDirection: {x: 0, y: 0}
+  dashUnlocked: 0
+  runUnlocked: 0
+  jumpUnlocked: 0
+  walkUnlocked: 0
 --- !u!212 &1051988535004059246
 SpriteRenderer:
   m_ObjectHideFlags: 0
@@ -169,12 +129,12 @@ SpriteRenderer:
   m_SortingLayerID: 0
   m_SortingLayer: 0
   m_SortingOrder: 0
-  m_Sprite: {fileID: -4398183398628465896, guid: d652faa8f10faa749af211cc5da6cf84, type: 3}
+  m_Sprite: {fileID: -4398183398628465896, guid: 4cb34ebff4c367f4e85ae9f08a95bfc2, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_FlipX: 0
   m_FlipY: 0
   m_DrawMode: 0
-  m_Size: {x: 1.0625, y: 2.125}
+  m_Size: {x: 0.23, y: 0.26}
   m_AdaptiveModeThreshold: 0.5
   m_SpriteTileMode: 0
   m_WasSpriteAssigned: 1
@@ -219,7 +179,7 @@ BoxCollider2D:
     border: {x: 0, y: 0, z: 0, w: 0}
     pivot: {x: 0.5, y: 0.5}
     oldSize: {x: 1.0625, y: 2.125}
-    newSize: {x: 1.0625, y: 2.125}
+    newSize: {x: 0.23, y: 0.26}
     adaptiveTilingThreshold: 0.5
     drawMode: 0
     adaptiveTiling: 0
@@ -275,37 +235,6 @@ Animator:
   m_AllowConstantClipSamplingOptimization: 1
   m_KeepAnimatorStateOnDisable: 0
   m_WriteDefaultValuesOnDisable: 0
---- !u!1 &1763808945923675895
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 8146929624418552425}
-  m_Layer: 2
-  m_Name: SlopeCheckFront
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!4 &8146929624418552425
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1763808945923675895}
-  serializedVersion: 2
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0.348, y: -1.061, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children: []
-  m_Father: {fileID: 7856122199337316976}
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &3858161665714943996
 GameObject:
   m_ObjectHideFlags: 0
