--- Assets/Scripts/ControllerPlayer.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 36bcc42383b3a8245b0d37a8d012a4bf
