--- Assets/Scripts/PlayerController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: eb98dd674770e234bac15bb300e91afb
